fix a bald trait preview bug (#6580)

+++ code/datums/preferences.dm
@@ -970,11 +970,16 @@ datum/preferences
 
 		src.preview?.update_appearance(src.AH, mutantRace, src.spessman_direction, name=src.real_name)
 
-		if (traitPreferences.traits_selected.Find("bald") && mutantRace)
-			var/mob/living/carbon/human/H = src.preview.preview_mob
-			var/ourWig = H.head
+		// bald trait preview stuff
+		if (!src.preview)
+			return
+		var/mob/living/carbon/human/H = src.preview.preview_mob
+		var/ourWig = H.head
+		if (ourWig)
 			H.u_equip(ourWig)
 			qdel(ourWig)
+
+		if (traitPreferences.traits_selected.Find("bald") && mutantRace)
 			H.equip_if_possible(H.create_wig(), H.slot_head)
 
 	proc/ShowChoices(mob/user)
